added error exceptions to desicion input function and tested it
